Better way to display secondary nav
The method I initially used was wp_list_pages(). The problem with wp_list_pages() is that it's page-based, not menu-based: the nav labels are the titles of the pages themselves, not the labels specified in the Menu panel.

Here's why this is a problem: in some cases, pages have very long titles, and as I realized, the vastly better way to handle these is to specify a different menu item label in the Menus area of WordPress—and not, as my last commit implied, set up some ungodly hack using custom fields. wp_list_pages() only knows the page titles; it has no idea if an alternate nav item title was specified in Menus. This commit displays the subnav as a menu instead of a list of pages; that will show whatever labels we've specified in the menu itself.

(I'm including the previous code; I'll be submitting a cleaned-up version of this shortly.)